refactor(session): move transcript shaping into session store
